requirements.txt
Django==1.6
Pillow==2.2.1
Pygments==1.6
South==0.8.4
Unipath==1.0
bpython==0.12
django-debug-toolbar==0.11.0
django-extensions==1.2.5
django-metadata==1.2
django-model-utils==1.5.0
django-mptt==0.6.0
six==1.4.1
sqlparse==0.1.10
psycopg2==2.5.1

# Heroku
dj-database-url==0.2.2
dj-static==0.0.5
gunicorn==18.0
static==0.4
wsgiref==0.1.2
